# docs/swagger.json
{
  "openapi":"3.1.0",
  "info":{
    "title":"yt-clipper-api",
    "summary":"Language-learning API for lessons, tutor, vocab, LINE, and AI workflows.",
    "description":"Language-learning API for YouTube lesson import, tutor sessions, vocabulary review, profile settings, AI speech utilities, and LINE-ready integration flows.\n\nAccess model: This backend runs in open-access compatibility mode. Requests do not need a prior login session. When no bearer token or auth cookies are provided, the service resolves a deterministic compatibility learner account so the API can still be called directly. Upstream gateways such as Kong can map real identity later.\n\nDocs:\n- Swagger UI: `/docs`\n- ReDoc: `/redoc`\n- OpenAPI JSON: `/openapi.json`",
    "version":"1.0.0"
  },
  "servers":[
    {
      "url":"http://localhost:8080",
      "description":"Configured public base URL"
    }
  ],
  "paths":{
...
    "/youtube/job/track":{
      "post":{
        "tags":[
          "youtube"
        ],
        "summary":"Track Legacy YouTube Job Progress",
        "description":"Update the last processed clip sequence for a legacy YouTube transaction job.",
        "operationId":"track_job_youtube_job_track_post",
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/JobTrackRequest"
              }
            }
          },
          "required":true
        },
        "responses":{
          "200":{
            "description":"Successful Response",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/YoutubeJobRes"
                }
              }
            }
          },
          "422":{
            "description":"Validation Error",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/youtube/jobs":{
      "post":{
        "tags":[
          "youtube"
        ],
        "summary":"Create Legacy YouTube Job",
        "description":"Create a raw YouTube processing job in the legacy pipeline and start transcript and clip generation asynchronously.",
        "operationId":"create_job_youtube_jobs_post",
        "requestBody":{
          "required":true,
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/CreateYoutubeJobReq"
              }
            }
          }
        },
        "responses":{
          "200":{
            "description":"Successful Response",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/YoutubeJobRes"
                }
              }
            }
          },
          "422":{
            "description":"Validation Error",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get":{
        "tags":[
          "youtube"
        ],
        "summary":"List Legacy YouTube Jobs",
        "description":"List legacy YouTube jobs by external user token for polling and history views.",
        "operationId":"list_jobs_youtube_jobs_get",
        "parameters":[
          {
            "name":"user_id_token",
            "in":"query",
            "required":true,
            "schema":{
              "type":"string",
              "title":"User Id Token"
            }
          }
        ],
        "responses":{
          "200":{
            "description":"Successful Response",
            "content":{
              "application/json":{
                "schema":{
                  "type":"array",
                  "items":{
                    "$ref":"#/components/schemas/YoutubeJobListItemRes"
                  },
                  "title":"Response List Jobs Youtube Jobs Get"
                }
              }
            }
          },
          "422":{
            "description":"Validation Error",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/youtube/jobs/{job_id}":{
      "get":{
        "tags":[
          "youtube"
        ],
        "summary":"Get Legacy YouTube Job Detail",
        "description":"Return a legacy YouTube job with processing progress, transcript source, clip artifacts, and per-segment details.",
        "operationId":"get_job_youtube_jobs__job_id__get",
        "parameters":[
          {
            "name":"job_id",
            "in":"path",
            "required":true,
            "schema":{
              "type":"integer",
              "title":"Job Id"
            }
          }
        ],
        "responses":{
          "200":{
            "description":"Successful Response",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/YoutubeJobDetailRes"
                }
              }
            }
          },
          "422":{
            "description":"Validation Error",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/youtube/translate":{
      "post":{
        "tags":[
          "youtube"
        ],
        "summary":"Translate Legacy YouTube Transcript",
        "description":"Translate transcript detail rows for a given `youtube_transaction_id` into `to_lang` using OpenAI `gpt-4o-mini`, then persist the result into `tbl_youtube_transaction_details.translate` and `to_lang`.",
        "operationId":"translate_job_youtube_translate_post",
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/TranslateYoutubeJobReq"
              }
            }
          },
          "required":true
        },
        "responses":{
          "200":{
            "description":"Successful Response",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/YoutubeTranslateRes"
                }
              }
            }
          },
          "422":{
            "description":"Validation Error",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components":{
    "schemas":{
...
      "ClipDetailRes":{
        "properties":{
          "id":{
            "type":"integer",
            "title":"Id"
          },
          "seq":{
            "type":"integer",
            "title":"Seq"
          },
          "start_s":{
            "type":"number",
            "title":"Start S"
          },
          "end_s":{
            "type":"number",
            "title":"End S"
          },
          "message":{
            "type":"string",
            "title":"Message"
          },
          "clip_path":{
            "anyOf":[
              {
                "type":"string"
              },
              {
                "type":"null"
              }
            ],
            "title":"Clip Path"
          },
          "has_clip":{
            "type":"boolean",
            "title":"Has Clip",
            "default":false
          },
          "url_video":{
            "anyOf":[
              {
                "type":"string"
              },
              {
                "type":"null"
              }
            ],
            "title":"Url Video"
          },
          "translate":{
            "anyOf":[
              {
                "type":"string"
              },
              {
                "type":"null"
              }
            ],
            "title":"Translate"
          },
          "to_lang":{
            "anyOf":[
              {
                "type":"string"
              },
              {
                "type":"null"
              }
            ],
            "title":"To Lang"
          }
        },
        "type":"object",
        "required":[
          "id",
          "seq",
          "start_s",
          "end_s",
          "message"
        ],
        "title":"ClipDetailRes"
      },
...
      "CreateYoutubeJobReq":{
        "properties":{
          "youtube_link":{
            "type":"string",
            "title":"Youtube Link"
          },
          "user_id_token":{
            "type":"string",
            "maxLength":36,
            "minLength":1,
            "title":"User Id Token"
          },
          "lang":{
            "type":"string",
            "title":"Lang",
            "default":"en"
          },
          "split_seconds":{
            "type":"integer",
            "title":"Split Seconds",
            "default":5
          },
          "tolerance_seconds":{
            "type":"integer",
            "title":"Tolerance Seconds",
            "default":1
          },
          "title":{
            "anyOf":[
              {
                "type":"string"
              },
              {
                "type":"null"
              }
            ],
            "title":"Title"
          },
          "type_of_transcription":{
            "anyOf":[
              {
                "type":"string"
              },
              {
                "type":"null"
              }
            ],
            "title":"Type Of Transcription"
          }
        },
        "type":"object",
        "required":[
          "youtube_link",
          "user_id_token"
        ],
        "title":"CreateYoutubeJobReq",
        "example":{
          "lang":"en",
          "split_seconds":15,
          "title":"Travel Listening Practice",
          "tolerance_seconds":1,
          "type_of_transcription":"openai_transcribe",
          "user_id_token":"device-123",
          "youtube_link":"https://www.youtube.com/watch?v=7IyiKqFQAnk"
        }
      },
...
      "HTTPValidationError":{
        "properties":{
          "detail":{
            "items":{
              "$ref":"#/components/schemas/ValidationError"
            },
            "type":"array",
            "title":"Detail"
          }
        },
        "type":"object",
        "title":"HTTPValidationError"
      },
...
      "JobTrackRequest":{
        "properties":{
          "job_id":{
            "type":"integer",
            "title":"Job Id"
          },
          "seq":{
            "type":"integer",
            "title":"Seq"
          }
        },
        "type":"object",
        "required":[
          "job_id",
          "seq"
        ],
        "title":"JobTrackRequest",
        "example":{
          "job_id":42,
          "seq":5
        }
      },
...
      "TranslateYoutubeJobReq":{
        "properties":{
          "youtube_transaction_id":{
            "type":"integer",
            "title":"Youtube Transaction Id"
          },
          "to_lang":{
            "type":"string",
            "maxLength":20,
            "minLength":2,
            "title":"To Lang"
          }
        },
        "type":"object",
        "required":[
          "youtube_transaction_id",
          "to_lang"
        ],
        "title":"TranslateYoutubeJobReq",
        "example":{
          "to_lang":"th",
          "youtube_transaction_id":42
        }
      },
...
      "ValidationError":{
        "properties":{
          "loc":{
            "items":{
              "anyOf":[
                {
                  "type":"string"
                },
                {
                  "type":"integer"
                }
              ]
            },
            "type":"array",
            "title":"Location"
          },
          "msg":{
            "type":"string",
            "title":"Message"
          },
          "type":{
            "type":"string",
            "title":"Error Type"
          }
        },
        "type":"object",
        "required":[
          "loc",
          "msg",
          "type"
        ],
        "title":"ValidationError"
      },
...
      "YoutubeJobDetailRes":{
        "properties":{
          "id":{
            "type":"integer",
            "title":"Id"
          },
          "youtube_link":{
            "type":"string",
            "title":"Youtube Link"
          },
          "title":{
            "anyOf":[
              {
                "type":"string"
              },
              {
                "type":"null"
              }
            ],
            "title":"Title"
          },
          "status":{
            "type":"string",
            "title":"Status"
          },
          "created_at":{
            "anyOf":[
              {
                "type":"string"
              },
              {
                "type":"null"
              }
            ],
            "title":"Created At"
          },
          "process_step":{
            "anyOf":[
              {
                "type":"string"
              },
              {
                "type":"null"
              }
            ],
            "title":"Process Step"
          },
          "progress":{
            "anyOf":[
              {
                "type":"integer"
              },
              {
                "type":"null"
              }
            ],
            "title":"Progress"
          },
          "source_transcript":{
            "anyOf":[
              {
                "type":"string"
              },
              {
                "type":"null"
              }
            ],
            "title":"Source Transcript"
          },
          "token_usage":{
            "anyOf":[
              {
                "type":"integer"
              },
              {
                "type":"null"
              }
            ],
            "title":"Token Usage"
          },
          "clip_error_code":{
            "anyOf":[
              {
                "type":"string"
              },
              {
                "type":"null"
              }
            ],
            "title":"Clip Error Code"
          },
          "clip_error_message":{
            "anyOf":[
              {
                "type":"string"
              },
              {
                "type":"null"
              }
            ],
            "title":"Clip Error Message"
          },
          "details":{
            "items":{
              "$ref":"#/components/schemas/ClipDetailRes"
            },
            "type":"array",
            "title":"Details"
          }
        },
        "type":"object",
        "required":[
          "id",
          "youtube_link",
          "title",
          "status",
          "details"
        ],
        "title":"YoutubeJobDetailRes"
      },
      "YoutubeJobListItemRes":{
        "properties":{
          "id":{
            "type":"integer",
            "title":"Id"
          },
          "created_at":{
            "anyOf":[
              {
                "type":"string"
              },
              {
                "type":"null"
              }
            ],
            "title":"Created At"
          },
          "title":{
            "anyOf":[
              {
                "type":"string"
              },
              {
                "type":"null"
              }
            ],
            "title":"Title"
          },
          "lastest_seq":{
            "anyOf":[
              {
                "type":"integer"
              },
              {
                "type":"null"
              }
            ],
            "title":"Lastest Seq"
          }
        },
        "type":"object",
        "required":[
          "id"
        ],
        "title":"YoutubeJobListItemRes"
      },
      "YoutubeJobRes":{
        "properties":{
          "id":{
            "type":"integer",
            "title":"Id"
          },
          "status":{
            "type":"string",
            "title":"Status"
          }
        },
        "type":"object",
        "required":[
          "id",
          "status"
        ],
        "title":"YoutubeJobRes"
      },
      "YoutubeTranslateRes":{
        "properties":{
          "youtube_transaction_id":{
            "type":"integer",
            "title":"Youtube Transaction Id"
          },
          "from_lang":{
            "anyOf":[
              {
                "type":"string"
              },
              {
                "type":"null"
              }
            ],
            "title":"From Lang"
          },
          "to_lang":{
            "type":"string",
            "title":"To Lang"
          },
          "translated_count":{
            "type":"integer",
            "title":"Translated Count"
          },
          "skipped_count":{
            "type":"integer",
            "title":"Skipped Count"
          },
          "details":{
            "items":{
              "$ref":"#/components/schemas/ClipDetailRes"
            },
            "type":"array",
            "title":"Details"
          }
        },
        "type":"object",
        "required":[
          "youtube_transaction_id",
          "to_lang",
          "translated_count",
          "skipped_count",
          "details"
        ],
        "title":"YoutubeTranslateRes"
      }
    }
  },
...
}